Initial check-in
A simple application that pops up to display the guimetadata responses
from the voice capabilities API. It is currently considered to be Alexa
specific, as it relies on the response templates provided by the Alexa
voiceagent via the capabilities API, and there are no other voiceagents
available to vet whether those template contents will be treated as a
fixed part of the capabilities API.  At the moment, only the two types
of body template (BodyTemplate1 and 2) and the WeatherTemplate are
handled, and rendering of the latter is only of the current day's
weather.

The application uses the onscreen role so that it can pop up above
running applications.  To drive the pop up behavior, a simple client is
used to handle the guimetadata events from vshl-capabilities and show
the window with the homescreen API.  Filling in the template is handled
with separate support from libqtappfw to provide the template data to
Qt.

Additionally, note that there currently is a hook for the navigation
capability in the simple afb client to enable raising the navigation
application when the setDestination capability is triggered.  Once a
better place to handle this is architected, it should be moved out of
this application.

Bug-AGL: SPEC-3110

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ifce9554b93d1e781283fd68f8e911a7010dd8cf0