app: Re-work the deactivation part in application_id_state event
In commit 11b9762f115778 we've added some logic in place that would
allow to start the application de-activated (that is hidden), as to
avoid 'stealing' the input focus (hoovering the pointer over a surface
will give that surface focus).

alexa-viewer application would be started by default by the system and
will be set 'invisible' (due to being a Qt app with the property set as
visible: false), and most likely this might cause a bit
of confusion for users. Also, by default, the compositor will display
the application as soon as started.

This patch re-works the de-activation part to handle a corner case, as
in commit 11b9762f115778 we would send the de-activate request as a
response in an event handler but it seems we can't guarantee that at that
time the surface was indeed activated.

So, instead of using application_id event, we use the application_id_state
event, which guarantees that we were actually sending the de-activate
request after we got the activate state event from the compositor.

Bug-AGL: SPEC-3412

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ie50b8011700d258a899374176fb78dcd528f88c1